+++ src/main/java/com/instaclick/pentaho/plugin/amqp/AMQPConfirmation.java
@@ -6,5 +6,4 @@ import java.io.IOException;
 public interface AMQPConfirmation {
     public void ackDelivery(long deliveryTag) throws IOException;
     public void rejectDelivery(long deliveryTag) throws IOException;
-    public void requeueDelivery(long deliveryTag) throws IOException;
 }